+++ History.md
@@ -1,3 +1,8 @@
+# 0.1.2 / 2014-02-10
+
+* Configuration files are now reloaded if `app-config` module is deleted from `require.cache` and then required again. This is useful for unit tests when need to reload the config files after making any changes to config object.
+* Replaced Node's `assert` module in unit tests by `should`
+
 # 0.1.1 / 2013-11-13
 
 * Always print an error when terminating the app due to a critical error